+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - quantityPerPage and quantityFirstPage on Reviews
+- `default ordination type` to the app's settings
 
 ## [1.2.2] - 2019-09-05
 ### Fixed